requirements.txt
# GitHub Issues & Projects 自動化システム 依存関係

# Gemini API
google-generativeai>=0.3.0

# GitHub API
PyGithub>=2.1.1

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Testing (optional)
pytest>=7.4.0
